+++ omaha/base/security/sha256.h
@@ -17,6 +17,7 @@
 #define OMAHA_BASE_SECURITY_SHA256_H_
 
 #include <stdint.h>
+#include <stddef.h>
 #include "hash-internal.h"
 
 #ifdef __cplusplus
@@ -26,11 +27,11 @@ extern "C" {
 typedef HASH_CTX LITE_SHA256_CTX;
 
 void SHA256_init(LITE_SHA256_CTX* ctx);
-void SHA256_update(LITE_SHA256_CTX* ctx, const void* data, unsigned int len);
+void SHA256_update(LITE_SHA256_CTX* ctx, const void* data, size_t len);
 const uint8_t* SHA256_final(LITE_SHA256_CTX* ctx);
 
 // Convenience method. Returns digest address.
-const uint8_t* SHA256_hash(const void* data, unsigned int len, uint8_t* digest);
+const uint8_t* SHA256_hash(const void* data, size_t len, uint8_t* digest);
 
 #define SHA256_DIGEST_SIZE 32
